+++ src/gridstack.ts
@@ -1,5 +1,5 @@
 /*!
- * GridStack 10.0.1-dev
+ * GridStack 10.1.2-dev
  * https://gridstackjs.com/
  *
  * Copyright (c) 2021-2022 Alain Dumesny
@@ -7,9 +7,11 @@
  */
 import { GridStackEngine } from './gridstack-engine';
 import { Utils, HeightData, obsolete, DragTransform } from './utils';
-import { gridDefaults, ColumnOptions, GridItemHTMLElement, GridStackElement, GridStackEventHandlerCallback,
+import {
+  gridDefaults, ColumnOptions, GridItemHTMLElement, GridStackElement, GridStackEventHandlerCallback,
   GridStackNode, GridStackWidget, numberOrString, DDUIData, DDDragInOpt, GridStackPosition, GridStackOptions,
-  dragInDefaultOptions, GridStackEventHandler, GridStackNodesHandler, AddRemoveFcn, SaveFcn, CompactOptions, GridStackMoveOpts, ResizeToContentFcn, GridStackDroppedHandler, GridStackElementHandler } from './types';
+  dragInDefaultOptions, GridStackEventHandler, GridStackNodesHandler, AddRemoveFcn, SaveFcn, CompactOptions, GridStackMoveOpts, ResizeToContentFcn, GridStackDroppedHandler, GridStackElementHandler
+} from './types';
 
 /*
  * and include D&D by default
@@ -99,7 +101,7 @@ export class GridStack {
     if (!el) {
       if (typeof elOrString === 'string') {
         console.error('GridStack.initAll() no grid was found with selector "' + elOrString + '" - element missing or wrong selector ?' +
-        '\nNote: ".grid-stack" is required for proper CSS styling and drag/drop, and is the default selector.');
+          '\nNote: ".grid-stack" is required for proper CSS styling and drag/drop, and is the default selector.');
       } else {
         console.error('GridStack.init() no grid element was passed.');
       }
@@ -131,7 +133,7 @@ export class GridStack {
     });
     if (grids.length === 0) {
       console.error('GridStack.initAll() no grid was found with selector "' + selector + '" - element missing or wrong selector ?' +
-      '\nNote: ".grid-stack" is required for proper CSS styling and drag/drop, and is the default selector.');
+        '\nNote: ".grid-stack" is required for proper CSS styling and drag/drop, and is the default selector.');
     }
     return grids;
   }
@@ -150,7 +152,7 @@ export class GridStack {
     if (el.gridstack) {
       // already a grid - set option and load data
       const grid = el.gridstack;
-      if (opt) grid.opts = {...grid.opts, ...opt};
+      if (opt) grid.opts = { ...grid.opts, ...opt };
       if (opt.children !== undefined) grid.load(opt.children);
       return grid;
     }
@@ -207,14 +209,8 @@ export class GridStack {
   /** scoping so users can call new GridStack.Engine(12) for example */
   public static Engine = GridStackEngine;
 
-  /** the HTML element tied to this grid after it's been initialized */
-  public el: GridHTMLElement;
-
   /** engine used to implement non DOM grid functionality */
   public engine: GridStackEngine;
-
-  /** grid options - public for classes to access, but use methods to modify! */
-  public opts: GridStackOptions;
 
   /** point to a parent grid item if we're nested (inside a grid-item in between 2 Grids) */
   public parentGridItem?: GridStackNode;
@@ -267,12 +263,11 @@ export class GridStack {
 
   /**
    * Construct a grid item from the given element and options
-   * @param el
-   * @param opts
+   * @param el the HTML element tied to this grid after it's been initialized
+   * @param opts grid options - public for classes to access, but use methods to modify!
    */
-  public constructor(el: GridHTMLElement, opts: GridStackOptions = {}) {
+  public constructor(public el: GridHTMLElement, public opts: GridStackOptions = {}) {
     el.gridstack = this;
-    this.el = el; // exposed HTML element to the user
     opts = opts || {}; // handles null/undefined/0
 
     if (!el.classList.contains('grid-stack')) {
@@ -309,8 +304,8 @@ export class GridStack {
       bk = opts.columnOpts.breakpoints = opts.columnOpts.breakpoints || [];
       let oneColumn = bk.find(b => b.c === 1);
       if (!oneColumn) {
-        oneColumn = {c: 1, w: oneSize};
-        bk.push(oneColumn, {c: 12, w: oneSize+1});
+        oneColumn = { c: 1, w: oneSize };
+        bk.push(oneColumn, { c: 12, w: oneSize + 1 });
       } else oneColumn.w = oneSize;
     }
     //...end LEGACY
@@ -324,10 +319,11 @@ export class GridStack {
         resp.columnMax = resp.columnMax || 12;
       }
     }
-    if (bk?.length > 1) bk.sort((a,b) => (b.w || 0) - (a.w || 0));
+    if (bk?.length > 1) bk.sort((a, b) => (b.w || 0) - (a.w || 0));
 
     // elements DOM attributes override any passed options (like CSS style) - merge the two together
-    let defaults: GridStackOptions = {...Utils.cloneDeep(gridDefaults),
+    let defaults: GridStackOptions = {
+      ...Utils.cloneDeep(gridDefaults),
       column: Utils.toNumber(el.getAttribute('gs-column')) || gridDefaults.column,
       minRow: rowAttr ? rowAttr : Utils.toNumber(el.getAttribute('gs-min-row')) || gridDefaults.minRow,
       maxRow: rowAttr ? rowAttr : Utils.toNumber(el.getAttribute('gs-max-row')) || gridDefaults.maxRow,
@@ -344,18 +340,17 @@ export class GridStack {
       defaults.animate = Utils.toBool(el.getAttribute('gs-animate'))
     }
 
-    this.opts = Utils.defaults(opts, defaults);
-    opts = null; // make sure we use this.opts instead
+    opts = Utils.defaults(opts, defaults);
     this._initMargin(); // part of settings defaults...
 
     // Now check if we're loading into 1 column mode FIRST so we don't do un-necessary work (like cellHeight = width / 12 then go 1 column)
     this.checkDynamicColumn();
-    this.el.classList.add('gs-' + this.opts.column);
+    this.el.classList.add('gs-' + opts.column);
 
-    if (this.opts.rtl === 'auto') {
-      this.opts.rtl = (el.style.direction === 'rtl');
+    if (opts.rtl === 'auto') {
+      opts.rtl = (el.style.direction === 'rtl');
     }
-    if (this.opts.rtl) {
+    if (opts.rtl) {
       this.el.classList.add('grid-stack-rtl');
     }
 
@@ -369,22 +364,22 @@ export class GridStack {
       parentGridItem.el.classList.add('grid-stack-sub-grid');
     }
 
-    this._isAutoCellHeight = (this.opts.cellHeight === 'auto');
-    if (this._isAutoCellHeight || this.opts.cellHeight === 'initial') {
+    this._isAutoCellHeight = (opts.cellHeight === 'auto');
+    if (this._isAutoCellHeight || opts.cellHeight === 'initial') {
       // make the cell content square initially (will use resize/column event to keep it square)
       this.cellHeight(undefined, false);
     } else {
       // append unit if any are set
-      if (typeof this.opts.cellHeight == 'number' && this.opts.cellHeightUnit && this.opts.cellHeightUnit !== gridDefaults.cellHeightUnit) {
-        this.opts.cellHeight = this.opts.cellHeight + this.opts.cellHeightUnit;
-        delete this.opts.cellHeightUnit;
+      if (typeof opts.cellHeight == 'number' && opts.cellHeightUnit && opts.cellHeightUnit !== gridDefaults.cellHeightUnit) {
+        opts.cellHeight = opts.cellHeight + opts.cellHeightUnit;
+        delete opts.cellHeightUnit;
       }
-      this.cellHeight(this.opts.cellHeight, false);
+      this.cellHeight(opts.cellHeight, false);
     }
 
     // see if we need to adjust auto-hide
-    if (this.opts.alwaysShowResizeHandle === 'mobile') {
-      this.opts.alwaysShowResizeHandle = isTouch;
+    if (opts.alwaysShowResizeHandle === 'mobile') {
+      opts.alwaysShowResizeHandle = isTouch;
     }
 
     this._styleSheetClass = 'gs-id-' + GridStackEngine._idSeq++;
@@ -392,11 +387,11 @@ export class GridStack {
 
     this._setStaticClass();
 
-    let engineClass = this.opts.engineClass || GridStack.engineClass || GridStackEngine;
+    let engineClass = opts.engineClass || GridStack.engineClass || GridStackEngine;
     this.engine = new engineClass({
       column: this.getColumn(),
-      float: this.opts.float,
-      maxRow: this.opts.maxRow,
+      float: opts.float,
+      maxRow: opts.maxRow,
       onChange: (cbNodes) => {
         let maxH = 0;
         this.engine.nodes.forEach(n => { maxH = Math.max(maxH, n.y + n.h) });
@@ -417,25 +412,25 @@ export class GridStack {
     // create initial global styles BEFORE loading children so resizeToContent margin can be calculated correctly
     this._updateStyles(false, 0);
 
-    if (this.opts.auto) {
+    if (opts.auto) {
       this.batchUpdate(); // prevent in between re-layout #1535 TODO: this only set float=true, need to prevent collision check...
       this.getGridItems().forEach(el => this._prepareElement(el));
       this.batchUpdate(false);
     }
 
     // load any passed in children as well, which overrides any DOM layout done above
-    if (this.opts.children) {
-      let children = this.opts.children;
-      delete this.opts.children;
+    if (opts.children) {
+      let children = opts.children;
+      delete opts.children;
       if (children.length) this.load(children); // don't load empty
     }
 
     // if (this.engine.nodes.length) this._updateStyles(); // update based on # of children. done in engine onChange CB
-    this.setAnimation(this.opts.animate);
+    this.setAnimation(opts.animate);
 
     // dynamic grids require pausing during drag to detect over to nest vs push
-    if (this.opts.subGridDynamic && !DDManager.pauseDrag) DDManager.pauseDrag = true;
-    if (this.opts.draggable?.pause !== undefined) DDManager.pauseDrag = this.opts.draggable.pause;
+    if (opts.subGridDynamic && !DDManager.pauseDrag) DDManager.pauseDrag = true;
+    if (opts.draggable?.pause !== undefined) DDManager.pauseDrag = opts.draggable.pause;
 
     this._setupRemoveDrop();
     this._setupAcceptWidget();
@@ -534,7 +529,7 @@ export class GridStack {
       grid = grid.parentGridItem?.grid;
     }
     //... and set the create options
-    ops = Utils.cloneDeep({...(subGridTemplate || {}), children: undefined, ...(ops || node.subGridOpts || {})});
+    ops = Utils.cloneDeep({ ...(subGridTemplate || {}), children: undefined, ...(ops || node.subGridOpts || {}) });
     node.subGridOpts = ops;
 
     // if column special case it set, remember that flag and set default
@@ -551,7 +546,7 @@ export class GridStack {
     let newItemOpt: GridStackNode;
     if (saveContent) {
       this._removeDD(node.el); // remove D&D since it's set on content div
-      newItemOpt = {...node, x:0, y:0};
+      newItemOpt = { ...node, x: 0, y: 0 };
       Utils.removeInternalForSave(newItemOpt);
       delete newItemOpt.subGridOpts;
       if (node.content) {
@@ -578,8 +573,8 @@ export class GridStack {
       let h = node.h + nodeToAdd.h;
       let style = node.el.style;
       style.transition = 'none'; // show up instantly so we don't see scrollbar with nodeToAdd
-      this.update(node.el, {w, h});
-      setTimeout(() =>  style.transition = null); // recover animation
+      this.update(node.el, { w, h });
+      setTimeout(() => style.transition = null); // recover animation
     }
 
     let subGrid = node.subGrid = GridStack.addGrid(content, ops);
@@ -653,7 +648,7 @@ export class GridStack {
         // check for nested grid
         if (n.subGrid?.el) {
           const listOrOpt = n.subGrid.save(saveContent, saveGridOpt, saveCB);
-          n.subGridOpts = (saveGridOpt ? listOrOpt : {children: listOrOpt}) as GridStackOptions;
+          n.subGridOpts = (saveGridOpt ? listOrOpt : { children: listOrOpt }) as GridStackOptions;
           delete n.subGrid;
         }
       }
@@ -704,9 +699,15 @@ export class GridStack {
     items = Utils.cloneDeep(items); // so we can mod
     const column = this.getColumn();
 
+    // if we have a mix of new items without coordinates and existing items, separate them out so they can be added after #2639
+    let addAfter = items.filter(n => (n.x === undefined || n.y === undefined) && !Utils.find(this.engine.nodes, n.id));
+    if (addAfter.length && addAfter.length !== items.length) {
+      items = items.filter(n => !Utils.find(addAfter, n.id));
+    } else addAfter = [];
+
     // if passed list has coordinates, use them (insert from end to beginning for conflict resolution) else keep widget order
     const haveCoord = items.some(w => w.x !== undefined || w.y !== undefined);
-    if (haveCoord) items = Utils.sort(items, -1, column);
+    if (haveCoord) items = Utils.sort(items, -1);
     this._insertNotAppend = haveCoord; // if we create in reverse order...
 
     // if we're loading a layout into for example 1 column and items don't fit, make sure to save
@@ -718,7 +719,7 @@ export class GridStack {
 
     // if given a different callback, temporally set it as global option so creating will use it
     const prevCB = GridStack.addRemoveCB;
-    if (typeof(addRemove) === 'function') GridStack.addRemoveCB = addRemove as AddRemoveFcn;
+    if (typeof (addRemove) === 'function') GridStack.addRemoveCB = addRemove as AddRemoveFcn;
 
     let removed: GridStackNode[] = [];
     this.batchUpdate();
@@ -765,7 +766,7 @@ export class GridStack {
         // add back to current list BUT force a collision check if it 'appears' we didn't change to make sure we don't overlap others now
         this.engine.nodes.push(item);
         if (Utils.samePos(item, w)) {
-          this.moveNode(item, {...w, forceCollide: true});
+          this.moveNode(item, { ...w, forceCollide: true });
         }
 
         this.update(item.el, w);
@@ -781,6 +782,11 @@ export class GridStack {
       }
     });
 
+    // finally append any separate ones that didn't have explicit coordinates last so they can find next empty spot
+    if (addRemove) {
+      addAfter.forEach(w => this.addWidget(w))
+    }
+
     this.engine.removedNodes = removed;
     this.batchUpdate(false);
 
@@ -788,8 +794,8 @@ export class GridStack {
     delete this._ignoreLayoutsNodeChange;
     delete this._insertNotAppend;
     prevCB ? GridStack.addRemoveCB = prevCB : delete GridStack.addRemoveCB;
-    // delay adding animation back
-    if (noAnim && this.opts.animate) setTimeout(() => this.setAnimation(this.opts.animate));
+    // delay adding animation back, but check to make sure grid (opt) is still around
+    if (noAnim && this.opts?.animate) setTimeout(() => { if (this.opts) this.setAnimation(this.opts.animate) });
     return this;
   }
 
@@ -813,7 +819,7 @@ export class GridStack {
    */
   public getCellHeight(forcePixel = false): number {
     if (this.opts.cellHeight && this.opts.cellHeight !== 'auto' &&
-       (!forcePixel || !this.opts.cellHeightUnit || this.opts.cellHeightUnit === 'px')) {
+      (!forcePixel || !this.opts.cellHeightUnit || this.opts.cellHeightUnit === 'px')) {
       return this.opts.cellHeight as number;
     }
     // do rem/em/cm/mm to px conversion
@@ -1041,12 +1047,12 @@ export class GridStack {
   public getCellFromPixel(position: MousePosition, useDocRelative = false): CellPosition {
     let box = this.el.getBoundingClientRect();
     // console.log(`getBoundingClientRect left: ${box.left} top: ${box.top} w: ${box.w} h: ${box.h}`)
-    let containerPos: {top: number, left: number};
+    let containerPos: { top: number, left: number };
     if (useDocRelative) {
-      containerPos = {top: box.top + document.documentElement.scrollTop, left: box.left};
+      containerPos = { top: box.top + document.documentElement.scrollTop, left: box.left };
       // console.log(`getCellFromPixel scrollTop: ${document.documentElement.scrollTop}`)
     } else {
-      containerPos = {top: this.el.offsetTop, left: this.el.offsetLeft}
+      containerPos = { top: this.el.offsetTop, left: this.el.offsetLeft }
       // console.log(`getCellFromPixel offsetTop: ${containerPos.left} offsetLeft: ${containerPos.top}`)
     }
     let relativeLeft = position.left - containerPos.left;
@@ -1055,7 +1061,7 @@ export class GridStack {
     let columnWidth = (box.width / this.getColumn());
     let rowHeight = (box.height / parseInt(this.el.getAttribute('gs-current-row')));
 
-    return {x: Math.floor(relativeLeft / columnWidth), y: Math.floor(relativeTop / rowHeight)};
+    return { x: Math.floor(relativeLeft / columnWidth), y: Math.floor(relativeTop / rowHeight) };
   }
 
   /** returns the current number of rows, which will be at least `minRow` if set */
@@ -1206,7 +1212,7 @@ export class GridStack {
       }
       if (!node) return;
 
-      if (GridStack.addRemoveCB) {
+      if (removeDOM && GridStack.addRemoveCB) {
         GridStack.addRemoveCB(this.el, node, false, false);
       }
 
@@ -1230,15 +1236,19 @@ export class GridStack {
   /**
    * Removes all widgets from the grid.
    * @param removeDOM if `false` DOM elements won't be removed from the tree (Default? `true`).
+   * @param triggerEvent if `false` (quiet mode) element will not be added to removed list and no 'removed' callbacks will be called (Default? true).
    */
-  public removeAll(removeDOM = true): GridStack {
+  public removeAll(removeDOM = true, triggerEvent = true): GridStack {
     // always remove our DOM data (circular link) before list gets emptied and drag&drop permanently
     this.engine.nodes.forEach(n => {
+      if (removeDOM && GridStack.addRemoveCB) {
+        GridStack.addRemoveCB(this.el, n, false, false);
+      }
       delete n.el.gridstackNode;
-      this._removeDD(n.el);
+      if (!this.opts.staticGrid) this._removeDD(n.el);
     });
-    this.engine.removeAll(removeDOM);
-    this._triggerRemoveEvent();
+    this.engine.removeAll(removeDOM, triggerEvent);
+    if (triggerEvent) this._triggerRemoveEvent();
     return this;
   }
 
@@ -1255,7 +1265,7 @@ export class GridStack {
     return this;
   }
   /** @internal */
-  private hasAnimationCSS(): boolean { return this.el.classList.contains('grid-stack-animate')  }
+  private hasAnimationCSS(): boolean { return this.el.classList.contains('grid-stack-animate') }
 
   /**
    * Toggle the grid static state, which permanently removes/add Drag&Drop support, unlike disable()/enable() that just turns it off/on.
@@ -1289,7 +1299,7 @@ export class GridStack {
       console.warn('gridstack.ts: `update(el, x, y, w, h)` is deprecated. Use `update(el, {x, w, content, ...})`. It will be removed soon');
       // eslint-disable-next-line prefer-rest-params
       let a = arguments, i = 1;
-      opt = { x:a[i++], y:a[i++], w:a[i++], h:a[i++] };
+      opt = { x: a[i++], y: a[i++], w: a[i++], h: a[i++] };
       return this.update(els, opt);
     }
 
@@ -1414,7 +1424,7 @@ export class GridStack {
     else if (n.maxH && h > n.maxH) h = n.maxH;
     if (h !== n.h) {
       grid._ignoreLayoutsNodeChange = true;
-      grid.moveNode(n, {h});
+      grid.moveNode(n, { h });
       delete grid._ignoreLayoutsNodeChange;
     }
   }
@@ -1467,7 +1477,7 @@ export class GridStack {
       console.warn('gridstack.ts: `willItFit(x,y,w,h,autoPosition)` is deprecated. Use `willItFit({x, y,...})`. It will be removed soon');
       // eslint-disable-next-line prefer-rest-params
       let a = arguments, i = 0,
-        w: GridStackWidget = { x:a[i++], y:a[i++], w:a[i++], h:a[i++], autoPosition:a[i++] };
+        w: GridStackWidget = { x: a[i++], y: a[i++], w: a[i++], h: a[i++], autoPosition: a[i++] };
       return this.willItFit(w);
     }
     return this.engine.willItFit(node);
@@ -1496,8 +1506,9 @@ export class GridStack {
       }
       // prevent added nodes from also triggering 'change' event (which is called next)
       this.engine.addedNodes.forEach(n => { delete n._dirty; });
-      this._triggerEvent('added', this.engine.addedNodes);
+      const addedNodes = [...this.engine.addedNodes];
       this.engine.addedNodes = [];
+      this._triggerEvent('added', addedNodes);
     }
     return this;
   }
@@ -1506,15 +1517,16 @@ export class GridStack {
   public _triggerRemoveEvent(): GridStack {
     if (this.engine.batchMode) return this;
     if (this.engine.removedNodes?.length) {
-      this._triggerEvent('removed', this.engine.removedNodes);
+      const removedNodes = [...this.engine.removedNodes];
       this.engine.removedNodes = [];
+      this._triggerEvent('removed', removedNodes);
     }
     return this;
   }
 
   /** @internal */
   protected _triggerEvent(type: string, data?: GridStackNode[]): GridStack {
-    let event = data ? new CustomEvent(type, {bubbles: false, detail: data}) : new Event(type);
+    let event = data ? new CustomEvent(type, { bubbles: false, detail: data }) : new Event(type);
     this.el.dispatchEvent(event);
     return this;
   }
@@ -1587,7 +1599,7 @@ export class GridStack {
       let getHeight = (rows: number): string => (cellHeight * rows) + cellHeightUnit;
       for (let i = this._styles._max + 1; i <= maxH; i++) { // start at 1
         Utils.addCSSRule(this._styles, `${prefix}[gs-y="${i}"]`, `top: ${getHeight(i)}`);
-        Utils.addCSSRule(this._styles, `${prefix}[gs-h="${i+1}"]`, `height: ${getHeight(i+1)}`); // start at 2
+        Utils.addCSSRule(this._styles, `${prefix}[gs-h="${i + 1}"]`, `height: ${getHeight(i + 1)}`); // start at 2
       }
       this._styles._max = maxH;
     }
@@ -1892,7 +1904,7 @@ export class GridStack {
     return this;
   }
 
-  static GDRev = '10.0.1-dev';
+  static GDRev = '10.1.2-dev';
 
   /* ===========================================================================================
    * drag&drop methods that used to be stubbed out and implemented in dd-gridstack.ts
@@ -1918,7 +1930,7 @@ export class GridStack {
       DDManager.pauseDrag = dragInOptions.pause;
     }
 
-    dragInOptions = {...dragInDefaultOptions, ...(dragInOptions || {})};
+    dragInOptions = { ...dragInDefaultOptions, ...(dragInOptions || {}) };
     let els: HTMLElement[] = (typeof dragIn === 'string') ? Utils.getElements(dragIn, root) : dragIn;
     if (els.length) els?.forEach(el => {
       if (!dd.isDraggable(el)) dd.dragIn(el, dragInOptions);
@@ -2061,7 +2073,7 @@ export class GridStack {
       }
 
       let parent = this.el.getBoundingClientRect();
-      let {top, left} = helper.getBoundingClientRect();
+      let { top, left } = helper.getBoundingClientRect();
       left -= parent.left;
       top -= parent.top;
       let ui: DDUIData = {
@@ -2101,7 +2113,7 @@ export class GridStack {
 
     dd.droppable(this.el, {
       accept: (el: GridItemHTMLElement) => {
-        let node: GridStackNode = el.gridstackNode;
+        let node: GridStackNode = el.gridstackNode || this._readAttr(el, false);
         // set accept drop to true on ourself (which we ignore) so we don't get "can't drop" icon in HTML5 mode while moving
         if (node?.grid === this) return true;
         if (!this.opts.acceptWidgets) return false;
@@ -2113,29 +2125,29 @@ export class GridStack {
           let selector = (this.opts.acceptWidgets === true ? '.grid-stack-item' : this.opts.acceptWidgets as string);
           canAccept = el.matches(selector);
         }
-        // finally check to make sure we actually have space left #1571
+        // finally check to make sure we actually have space left #1571 #2633
         if (canAccept && node && this.opts.maxRow) {
-          let n = {w: node.w, h: node.h, minW: node.minW, minH: node.minH}; // only width/height matters and autoPosition
+          let n = { w: node.w, h: node.h, minW: node.minW, minH: node.minH }; // only width/height matters and autoPosition
           canAccept = this.engine.willItFit(n);
         }
         return canAccept;
       }
     })
-    /**
-     * entering our grid area
-     */
+      /**
+       * entering our grid area
+       */
       .on(this.el, 'dropover', (event: Event, el: GridItemHTMLElement, helper: GridItemHTMLElement) => {
-      // console.log(`over ${this.el.gridstack.opts.id} ${count++}`); // TEST
+        // console.log(`over ${this.el.gridstack.opts.id} ${count++}`); // TEST
         let node = el.gridstackNode;
         // ignore drop enter on ourself (unless we temporarily removed) which happens on a simple drag of our item
         if (node?.grid === this && !node._temporaryRemoved) {
-        // delete node._added; // reset this to track placeholder again in case we were over other grid #1484 (dropout doesn't always clear)
+          // delete node._added; // reset this to track placeholder again in case we were over other grid #1484 (dropout doesn't always clear)
           return false; // prevent parent from receiving msg (which may be a grid as well)
         }
 
         // fix #1578 when dragging fast, we may not get a leave on the previous grid so force one now
         if (node?.grid && node.grid !== this && !node._temporaryRemoved) {
-        // console.log('dropover without leave'); // TEST
+          // console.log('dropover without leave'); // TEST
           let otherGrid = node.grid;
           otherGrid._leave(el, helper);
         }
@@ -2160,36 +2172,36 @@ export class GridStack {
 
         // if the item came from another grid, make a copy and save the original info in case we go back there
         if (node.grid && node.grid !== this) {
-        // copy the node original values (min/max/id/etc...) but override width/height/other flags which are this grid specific
-        // console.log('dropover cloning node'); // TEST
+          // copy the node original values (min/max/id/etc...) but override width/height/other flags which are this grid specific
+          // console.log('dropover cloning node'); // TEST
           if (!el._gridstackNodeOrig) el._gridstackNodeOrig = node; // shouldn't have multiple nested!
-          el.gridstackNode = node = {...node, w, h, grid: this};
+          el.gridstackNode = node = { ...node, w, h, grid: this };
           delete node.x;
           delete node.y;
           this.engine.cleanupNode(node)
             .nodeBoundFix(node);
           // restore some internal fields we need after clearing them all
           node._initDD =
-          node._isExternal =  // DOM needs to be re-parented on a drop
-          node._temporaryRemoved = true; // so it can be inserted onDrag below
+            node._isExternal =  // DOM needs to be re-parented on a drop
+            node._temporaryRemoved = true; // so it can be inserted onDrag below
         } else {
           node.w = w; node.h = h;
           node._temporaryRemoved = true; // so we can insert it
         }
 
         // clear any marked for complete removal (Note: don't check _isAboutToRemove as that is cleared above - just do it)
-        this._itemRemoving(node.el, false);
+        GridStack._itemRemoving(node.el, false);
 
         dd.on(el, 'drag', onDrag);
         // make sure this is called at least once when going fast #1578
         onDrag(event as DragEvent, el, helper);
         return false; // prevent parent from receiving msg (which may be a grid as well)
       })
-    /**
-     * Leaving our grid area...
-     */
+      /**
+       * Leaving our grid area...
+       */
       .on(this.el, 'dropout', (event, el: GridItemHTMLElement, helper: GridItemHTMLElement) => {
-      // console.log(`out ${this.el.gridstack.opts.id} ${count++}`); // TEST
+        // console.log(`out ${this.el.gridstack.opts.id} ${count++}`); // TEST
         let node = el.gridstackNode;
         if (!node) return false;
         // fix #1578 when dragging fast, we might get leave after other grid gets enter (which calls us to clean)
@@ -2203,9 +2215,9 @@ export class GridStack {
         }
         return false; // prevent parent from receiving msg (which may be grid as well)
       })
-    /**
-     * end - releasing the mouse
-     */
+      /**
+       * end - releasing the mouse
+       */
       .on(this.el, 'drop', (event, el: GridItemHTMLElement, helper: GridItemHTMLElement) => {
         let node = el.gridstackNode;
         // ignore drop on ourself from ourself that didn't come from the outside - dragend will handle the simple move instead
@@ -2279,11 +2291,11 @@ export class GridStack {
 
         this.engine.endUpdate();
         if (this._gsEventHandler['dropped']) {
-          this._gsEventHandler['dropped']({...event, type: 'dropped'}, origNode && origNode.grid ? origNode : undefined, node);
+          this._gsEventHandler['dropped']({ ...event, type: 'dropped' }, origNode && origNode.grid ? origNode : undefined, node);
         }
 
-        // delay adding animation back
-        if (noAnim) setTimeout(() => this.setAnimation(this.opts.animate));
+        // delay adding animation back, but check to make sure grid (opt) is still around
+        if (noAnim) setTimeout(() => { if (this.opts) this.setAnimation(this.opts.animate) });
 
         return false; // prevent parent from receiving msg (which may be grid as well)
       });
@@ -2291,26 +2303,26 @@ export class GridStack {
   }
 
   /** @internal mark item for removal */
-  private _itemRemoving(el: GridItemHTMLElement, remove: boolean) {
-    let node = el ? el.gridstackNode : undefined;
-    if (!node || !node.grid || el.classList.contains(this.opts.removableOptions.decline)) return;
+  private static _itemRemoving(el: GridItemHTMLElement, remove: boolean) {
+    const node = el ? el.gridstackNode : undefined;
+    if (!node?.grid || el.classList.contains(node.grid.opts.removableOptions.decline)) return;
     remove ? node._isAboutToRemove = true : delete node._isAboutToRemove;
     remove ? el.classList.add('grid-stack-item-removing') : el.classList.remove('grid-stack-item-removing');
   }
 
   /** @internal called to setup a trash drop zone if the user specifies it */
   protected _setupRemoveDrop(): GridStack {
-    if (!this.opts.staticGrid && typeof this.opts.removable === 'string') {
-      let trashEl = document.querySelector(this.opts.removable) as HTMLElement;
-      if (!trashEl) return this;
-      // only register ONE drop-over/dropout callback for the 'trash', and it will
-      // update the passed in item and parent grid because the 'trash' is a shared resource anyway,
-      // and Native DD only has 1 event CB (having a list and technically a per grid removableOptions complicates things greatly)
-      if (!dd.isDroppable(trashEl)) {
-        dd.droppable(trashEl, this.opts.removableOptions)
-          .on(trashEl, 'dropover', (event, el) => this._itemRemoving(el, true))
-          .on(trashEl, 'dropout',  (event, el) => this._itemRemoving(el, false));
-      }
+    if (typeof this.opts.removable !== 'string') return this;
+    let trashEl = document.querySelector(this.opts.removable) as HTMLElement;
+    if (!trashEl) return this;
+
+    // only register ONE static drop-over/dropout callback for the 'trash', and it will
+    // update the passed in item and parent grid because the '.trash' is a shared resource anyway,
+    // and Native DD only has 1 event CB (having a list and technically a per grid removableOptions complicates things greatly)
+    if (!this.opts.staticGrid && !dd.isDroppable(trashEl)) {
+      dd.droppable(trashEl, this.opts.removableOptions)
+        .on(trashEl, 'dropover', (event, el) => GridStack._itemRemoving(el, true))
+        .on(trashEl, 'dropout', (event, el) => GridStack._itemRemoving(el, false));
     }
     return this;
   }
@@ -2475,7 +2487,7 @@ export class GridStack {
 
   /** @internal handles actual drag/resize */
   protected _dragOrResize(el: GridItemHTMLElement, event: MouseEvent, ui: DDUIData, node: GridStackNode, cellWidth: number, cellHeight: number): void {
-    let p = {...node._orig}; // could be undefined (_isExternal) which is ok (drag only set x,y and w,h will default to node value)
+    let p = { ...node._orig }; // could be undefined (_isExternal) which is ok (drag only set x,y and w,h will default to node value)
     let resizing: boolean;
     let mLeft = this.opts.marginLeft as number,
       mRight = this.opts.marginRight as number,
@@ -2499,8 +2511,8 @@ export class GridStack {
       }
 
       // get new position taking into account the margin in the direction we are moving! (need to pass mid point by margin)
-      let left = ui.position.left + (ui.position.left > node._lastUiPosition.left  ? -mRight : mLeft);
-      let top = ui.position.top + (ui.position.top > node._lastUiPosition.top  ? -mBottom : mTop);
+      let left = ui.position.left + (ui.position.left > node._lastUiPosition.left ? -mRight : mLeft);
+      let top = ui.position.top + (ui.position.top > node._lastUiPosition.top ? -mBottom : mTop);
       p.x = Math.round(left / cellWidth);
       p.y = Math.round(top / cellHeight);
 
@@ -2519,7 +2531,7 @@ export class GridStack {
       if (node.x === p.x && node.y === p.y) return; // skip same
       // DON'T skip one we tried as we might have failed because of coverage <50% before
       // if (node._lastTried && node._lastTried.x === x && node._lastTried.y === y) return;
-    } else if (event.type === 'resize')  {
+    } else if (event.type === 'resize') {
       if (p.x < 0) return;
       // Scrolling page if needed
       Utils.updateScrollResize(event, el, cellHeight);
@@ -2547,7 +2559,7 @@ export class GridStack {
       w: (ui.size ? ui.size.width : node.w * cellWidth) - mLeft - mRight,
       h: (ui.size ? ui.size.height : node.h * cellHeight) - mTop - mBottom
     };
-    if (this.engine.moveNodeCheck(node, {...p, cellWidth, cellHeight, rect, resizing})) {
+    if (this.engine.moveNodeCheck(node, { ...p, cellWidth, cellHeight, rect, resizing })) {
       node._lastUiPosition = ui.position;
       this.engine.cacheRects(cellWidth, cellHeight, mTop, mRight, mBottom, mLeft);
       delete node._skipDown;
@@ -2585,7 +2597,7 @@ export class GridStack {
 
     if (this.opts.removable === true) { // boolean vs a class string
       // item leaving us and we are supposed to remove on leave (no need to drag onto trash) mark it so
-      this._itemRemoving(el, true);
+      GridStack._itemRemoving(el, true);
     }
 
     // finally if item originally came from another grid, but left us, restore things back to prev info